Added repair method to player inventory and tower

+++ src/main/java/seng201/team0/models/Consumables/Booster.java
@@ -3,7 +3,7 @@ package seng201.team0.models.Consumables;
 import seng201.team0.models.Item;
 
 public abstract class Booster extends Item {
-    public Booster(String inputName, String inputDescription, double inputCost) {
+    public Booster(String inputName, String inputDescription, int inputCost) {
         super(inputName, inputDescription, inputCost);
     }
     public abstract void apply();
